% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.tscount-package.R
\docType{package}
\name{fable.tscount-package}
\alias{fable.tscount}
\alias{fable.tscount-package}
\title{fable.tscount: Fable Modelling Wrappers for 'tscount'}
\description{
Allows tsglm models from the 'tscount' package to
    be used in a tidy workflow with the modelling interface of
    'fabletools'. This extends 'tscount' to provide enhanced model
    specification and management, performance evaluation methods, and
    model combination tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mitchelloharawild/fable.tscount}
  \item Report bugs at \url{https://github.com/mitchelloharawild/fable.tscount/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

}
\keyword{internal}
